+++ packages/thirdweb/src/react/web/ui/ConnectWallet/screens/Buy/main/BuyUIMainScreen.tsx
@@ -17,8 +17,8 @@ import { Button } from "../../../../components/buttons.js";
 import { Text } from "../../../../components/text.js";
 import type { PayUIOptions } from "../../../ConnectButtonProps.js";
 import { type ERC20OrNativeToken, NATIVE_TOKEN } from "../../nativeToken.js";
+import type { SupportedChainAndTokens } from "../fetchChainsAndTokens.js";
 import { BuyTokenInput } from "../swap/BuyTokenInput.js";
-import type { SupportedChainAndTokens } from "../swap/useSwapSupportedChains.js";
 import type { BuyForTx, SelectedScreen } from "./types.js";
 import { useBuyTxStates } from "./useBuyTxStates.js";
 import { useEnabledPaymentMethods } from "./useEnabledPaymentMethods.js";
